Move RpcService into its own module
