+++ src/lib.rs
@@ -63,6 +63,7 @@ pub mod rpc_mock;
 pub mod rpc_pubsub;
 pub mod rpc_pubsub_service;
 pub mod rpc_request;
+pub mod rpc_service;
 pub mod rpc_subscriptions;
 pub mod service;
 pub mod sigverify;
